I have done changes concerning comments.

+++ os-server/src/main/java/app/openschool/course/CourseController.java
@@ -6,8 +6,6 @@ import app.openschool.course.api.dto.CourseInfoDto;
 import app.openschool.course.api.dto.CreateCourseRequest;
 import app.openschool.course.api.dto.UpdateCourseRequest;
 import app.openschool.course.api.mapper.CourseMapper;
-import app.openschool.user.User;
-import app.openschool.user.UserRepository;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -17,14 +15,12 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import java.security.Principal;
 import java.util.List;
-import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -40,16 +36,9 @@ import org.springframework.web.bind.annotation.RestController;
 public class CourseController {
 
   private final CourseService courseService;
-  private final UserRepository userRepository;
-  private final CourseRepository courseRepository;
 
-  public CourseController(
-      CourseService courseService,
-      UserRepository userRepository,
-      CourseRepository courseRepository) {
+  public CourseController(CourseService courseService) {
     this.courseService = courseService;
-    this.userRepository = userRepository;
-    this.courseRepository = courseRepository;
   }
 
   @Operation(summary = "get course info", security = @SecurityRequirement(name = "bearerAuth"))
@@ -156,9 +145,8 @@ public class CourseController {
           @RequestBody
           UpdateCourseRequest request,
       Principal principal) {
-    Course courseById =
-        courseRepository.findById(courseId).orElseThrow(IllegalArgumentException::new);
-    if (!principal.getName().equals(courseById.getMentor().getEmail())) {
+    Course course = courseService.findCourseById(courseId).orElseThrow(IllegalAccessError::new);
+    if (!principal.getName().equals(course.getMentor().getEmail())) {
       throw new IllegalArgumentException();
     }
     return ResponseEntity.ok()
@@ -185,7 +173,13 @@ public class CourseController {
   @DeleteMapping("/{id}")
   public ResponseEntity<HttpStatus> delete(
       @Parameter(description = "Id of course which will be deleted") @PathVariable(value = "id")
-          Long courseId) {
+          Long courseId,
+      Principal principal) {
+    Course course =
+        courseService.findCourseById(courseId).orElseThrow(IllegalArgumentException::new);
+    if (!course.getMentor().getEmail().equals(principal.getName())) {
+      throw new IllegalArgumentException();
+    }
     courseService.delete(courseId);
     return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
   }
